Refactor Linux theme detection
