+++ src/qz/installer/LinuxInstaller.java
@@ -5,6 +5,7 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import qz.App;
 import qz.utils.*;
+import qz.utils.linux.Distro;
 
 import javax.swing.*;
 import java.awt.*;
@@ -110,7 +111,7 @@ public class LinuxInstaller extends Installer {
                                   "cups-bsd", "cups-client", "cups");
 
         // Legacy Ubuntu versions only: Patch Unity to show the System Tray
-        if(UnixUtilities.isUbuntu()) {
+        if(Distro.isUbuntu()) {
             ShellUtilities.execute("gsettings", "set", "com.canonical.Unity.Panel", "systray", "-whitelist", "\"['all']\"");
 
             if(ShellUtilities.execute("killall", "-w", "unity", "-panel")) {
@@ -380,11 +381,11 @@ public class LinuxInstaller extends Installer {
         boolean installed = ShellUtilities.execute("which", binary);
         if (!installed) {
             if (SystemUtilities.isAdmin() && promptInstall(binary)) {
-                if(UnixUtilities.isDebian()) {
+                if(Distro.isDebian()) {
                     installed = ShellUtilities.execute("apt-get", "install", "-y", debianPackage);
-                } else if(UnixUtilities.isFedora()) {
+                } else if(Distro.isFedora()) {
                     installed =  ShellUtilities.execute("dnf", "install", "-y", fedoraPackage);
-                } else if(UnixUtilities.isArch()) {
+                } else if(Distro.isArch()) {
                     installed =  ShellUtilities.execute("pacman", "-S", "--noconfirm", archPackage);
                 }
             }
