+++ src/qz/utils/SystemUtilities.java
@@ -25,7 +25,9 @@ import qz.build.provision.params.Arch;
 import qz.build.provision.params.Os;
 import qz.common.Constants;
 import qz.installer.Installer;
-import qz.utils.gtk.GtkUtilities;
+import qz.utils.linux.DeUtils;
+import qz.utils.linux.Distro;
+import qz.utils.linux.gtk.GtkScale;
 
 import javax.swing.*;
 import java.awt.*;
@@ -157,17 +159,9 @@ public class SystemUtilities {
     public static String getOsDisplayVersion() {
         if (osDisplayVersion == null) {
             switch(OS_TYPE) {
-                case WINDOWS:
-                    osDisplayVersion = WindowsUtilities.getOsDisplayVersion();
-                    break;
-                case MAC:
-                    osDisplayVersion = MacUtilities.getOsDisplayVersion();
-                    break;
-                case LINUX:
-                    osDisplayVersion = UnixUtilities.getOsDisplayVersion();
-                    break;
-                default:
-                    osDisplayVersion = System.getProperty("os.version", "0.0.0");
+                case WINDOWS -> osDisplayVersion = WindowsUtilities.getOsDisplayVersion();
+                case MAC -> osDisplayVersion = MacUtilities.getOsDisplayVersion();
+                default -> osDisplayVersion = Distro.getOsDisplayVersion();
             }
         }
         return osDisplayVersion;
@@ -179,12 +173,8 @@ public class SystemUtilities {
     public static String getOsDisplayName() {
         if(osName == null) {
             switch(OS_TYPE) {
-                case LINUX:
-                    // "Linux" is too generic, get the flavor (e.g. Ubuntu, Fedora)
-                    osName = UnixUtilities.getDisplayName();
-                    break;
-                default:
-                    osName = System.getProperty("os.name", "Unknown");
+                case WINDOWS, MAC -> osName = System.getProperty("os.name", "Unknown");
+                default -> osName = Distro.getDisplayName();
             }
         }
         return osName;
@@ -326,19 +316,20 @@ public class SystemUtilities {
         return isDarkTaskbar(false);
     }
 
+    /**
+     * Handle nuances of a taskbar that's themed differently than the rest of the desktop
+     * which is mostly a Windows-ism.  macOS and Linux both provide ways to use template
+     * icons/mask icons for the taskbar.
+     */
     public static boolean isDarkTaskbar(boolean recheck) {
         if(darkTaskbar == null || recheck) {
-            if (isWindows()) {
-                darkTaskbar = WindowsUtilities.isDarkTaskbar();
-            } else if(isMac()) {
-                // Ignore, we'll set the template flag using JNA
-                darkTaskbar = false;
-            } else {
-                // Linux doesn't differentiate; return the cached darkDesktop value
-                darkTaskbar = isDarkDesktop();
-            }
+            darkTaskbar = switch(OS_TYPE) {
+                case WINDOWS -> WindowsUtilities.isDarkTaskbar();
+                case MAC -> false; // ignore, we'll use template icons
+                default ->  isDarkDesktop(); // re-use desktop value TODO: revisit after #1456
+            };
         }
-        return darkTaskbar.booleanValue();
+        return darkTaskbar;
     }
 
     public static boolean isDarkDesktop() {
@@ -347,16 +338,13 @@ public class SystemUtilities {
 
     public static boolean isDarkDesktop(boolean recheck) {
         if (darkDesktop == null || recheck) {
-            // Check for Dark Mode on MacOS
-            if (isMac()) {
-                darkDesktop = MacUtilities.isDarkDesktop();
-            } else if (isWindows()) {
-                darkDesktop = WindowsUtilities.isDarkDesktop();
-            } else {
-                darkDesktop = UnixUtilities.isDarkDesktop();
-            }
+            darkDesktop = switch(OS_TYPE) {
+                case MAC -> MacUtilities.isDarkDesktop();
+                case WINDOWS -> WindowsUtilities.isDarkDesktop();
+                default ->  DeUtils.isDarkDesktop();
+            };
         }
-        return darkDesktop.booleanValue();
+        return darkDesktop;
     }
 
     public static void adjustThemeColors() {
@@ -365,15 +353,11 @@ public class SystemUtilities {
     }
 
     public static boolean prefersMaskTrayIcon() {
-        if (Constants.MASK_TRAY_SUPPORTED) {
-            if (SystemUtilities.isMac()) {
-                // Assume a pid of -1 is a broken JNA
-                return getProcessId() != -1;
-            } else if (SystemUtilities.isWindows() && SystemUtilities.getOsVersion().getMajorVersion() >= 10) {
-                return true;
-            }
-        }
-        return false;
+        return Constants.MASK_TRAY_SUPPORTED && switch(OS_TYPE) {
+            case MAC -> true;
+            case WINDOWS -> SystemUtilities.getOsVersion().majorVersion() >= 10;
+            default -> false;
+        };
     }
 
     /**
@@ -385,7 +369,7 @@ public class SystemUtilities {
     private static String calculateLightLaf() {
         return switch(OS_TYPE) {
             case WINDOWS, MAC -> UIManager.getSystemLookAndFeelClassName();
-            default -> GtkUtilities.isGtkAvailable()?
+            default -> GtkScale.isGtkAvailable()?
                     "com.sun.java.swing.plaf.gtk.GTKLookAndFeel":
                     FlatIntelliJLaf.class.getCanonicalName();
         };
@@ -397,7 +381,7 @@ public class SystemUtilities {
      * fall back on a suitable dark-mode theme.
      */
     private static String calculateDarkLaf() {
-        return GtkUtilities.isGtkAvailable() ?
+        return GtkScale.isGtkAvailable() ?
                 "com.sun.java.swing.plaf.gtk.GTKLookAndFeel" :
                 FlatDarculaLaf.class.getCanonicalName();
     }
@@ -479,20 +463,10 @@ public class SystemUtilities {
      */
     private static double getWindowScaleFactor(boolean forceRefresh) {
         if(windowScaleFactor == -1 || forceRefresh) {
-            // MacOS is always 1
-            if (isMac()) {
-                return windowScaleFactor = 1;
+            switch(OS_TYPE) {
+                case MAC, WINDOWS -> windowScaleFactor = 1;
+                default -> windowScaleFactor = GtkScale.getScaleFactor();
             }
-            // Windows/Linux on JDK8 honors scaling
-            if (Constants.JAVA_VERSION.lessThan(Version.valueOf("11.0.0"))) {
-                return windowScaleFactor = Toolkit.getDefaultToolkit().getScreenResolution() / 96.0;
-            }
-            // Windows on JDK11 is always 1
-            if (isWindows()) {
-                return windowScaleFactor = 1;
-            }
-            // Linux/Unix on JDK11 requires JNA calls to Gdk
-            return windowScaleFactor = UnixUtilities.getScaleFactor();
         }
         return windowScaleFactor;
     }
@@ -511,25 +485,6 @@ public class SystemUtilities {
                 (int)(width * scaleFactor),
                 (int)(height * scaleFactor)
         );
-    }
-
-    /**
-     * Detects if HiDPI is enabled
-     * Warning: Due to behavioral differences between OSs, JDKs and poor
-     * detection techniques this function should only be used to fix rare
-     * edge-case bugs.
-     *
-     * See also SystemUtilities.getWindowScaleFactor()
-     * @return true if HiDPI is detected
-     */
-    public static boolean isHiDPI() {
-        if(isMac()) {
-            return MacUtilities.getScaleFactor() > 1;
-        } else if(isWindows()) {
-            return WindowsUtilities.getScaleFactor() > 1;
-        }
-        // Fallback to a JNA Gdk technique
-        return UnixUtilities.getScaleFactor() > 1;
     }
 
     /**
